webapp/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
supabase>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
imagehash>=4.3.0
Pillow>=9.5.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
ultralytics>=8.0.0
pytesseract>=0.3.10
gunicorn>=21.0.0 